:zap: fix performance issue with every chunk size configuration
close https://github.com/jawah/niquests/issues/236
